+++ CHANGELOG.md
@@ -21,6 +21,9 @@ and this project adheres to
     datacenters, nodes, racks and infrastructures subcommands.
   - Add possibility to select output format of the list of datacenters,
     infrastructures, nodes and racks names.
+- lib:
+  - Add RacksDB.racks property to get the full list of racks.
+  - Add possibility to filter list DatacenterRoomRack by name.
 - docs:
   - Add installation method from sources intended to software developers in the
     quickstart guide.